Added Webpack for file splitting
