+++ src/Darkys_Achievement_Package/Achievement/index.js
@@ -1,0 +1,8 @@
+const Achievement = (name, desc, icon) => {
+    const answer = new Game.Achievement(name, desc, icon);
+    answer.darky = 1;
+
+    return answer;
+};
+
+export default Achievement;
