+++ src/Darkys_Achievement_Package/init/checkHook/components/bigCookieClicks/index.js
@@ -1,0 +1,19 @@
+const bigCookieClicks = [
+    () => {
+        if (Game.cookieClicks >= 1000) {
+            Game.Win("Clicker");
+        }
+    },
+    () => {
+        if (Game.cookieClicks >= 10000) {
+            Game.Win("Advanced Clicker");
+        }
+    },
+    () => {
+        if (Game.cookieClicks >= 25000) {
+            Game.Win("Expert Clicker");
+        }
+    },
+];
+
+export default bigCookieClicks;
