+++ src/Darkys_Achievement_Package/init/checkHook/components/cookiesFromClicking/index.js
@@ -1,0 +1,58 @@
+const cookiesFromClicking = [
+    () => {
+        if (Game.handmadeCookies >= 1e29) {
+            Game.Win("Stop it, it clickles");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e31) {
+            Game.Win("Keep up the click");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e33) {
+            Game.Win("Clicking is the new booping");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e35) {
+            Game.Win("The final click");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e37) {
+            Game.Win("Sike, thats the wrong achievement");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e39) {
+            Game.Win("How about we tap the cookie now");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e41) {
+            Game.Win("Cookie Destroyer");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e43) {
+            Game.Win("Give it a click");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e45) {
+            Game.Win("Clicking Frenzy: Our Lord and Saviour");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e47) {
+            Game.Win("Hand-made cookies are still the best");
+        }
+    },
+    () => {
+        if (Game.handmadeCookies >= 1e49) {
+            Game.Win("My strange clicking addiction");
+        }
+    },
+];
+export default cookiesFromClicking;
