+++ src/Darkys_Achievement_Package/init/checkHook/components/index.js
@@ -1,0 +1,11 @@
+export { default as bigCookieClicks } from "./bigCookieClicks";
+export { default as buildingsLevel } from "./buildingsLevel";
+export { default as buildingsOwned } from "./buildingsOwned";
+export { default as cookiesFromClicking } from "./cookiesFromClicking";
+export { default as everyBuildingOwned } from "./everyBuildingOwned";
+export { default as prestigeAmount } from "./prestigeAmount";
+export { default as prestigeUpgradesOwned } from "./prestigeUpgradesOwned";
+export { default as shadow } from "./shadow";
+export { default as themedUpgrades } from "./themedUpgrades";
+export { default as totalBuildingsOwned } from "./totalBuildingsOwned";
+export { default as totalUpgradesOwned } from "./totalUpgradesOwned";
