+++ src/Darkys_Achievement_Package/init/checkHook/components/shadow/index.js
@@ -1,0 +1,13 @@
+const shadow = [
+    () => {
+        if (Game.bakeryName === "Darky") {
+            Game.Win("Mod-God complex");
+        }
+    },
+    () => {
+        if (Game.goldenClicks >= 77777) {
+            Game.Win("Golden Jackpot");
+        }
+    },
+];
+export default shadow;
