+++ src/Darkys_Achievement_Package/init/checkHook/index.js
@@ -1,0 +1,29 @@
+import {
+    bigCookieClicks,
+    buildingsLevel,
+    buildingsOwned,
+    cookiesFromClicking,
+    everyBuildingOwned,
+    prestigeAmount,
+    prestigeUpgradesOwned,
+    shadow,
+    themedUpgrades,
+    totalBuildingsOwned,
+    totalUpgradesOwned,
+} from "./components";
+
+const checkHookContent = [
+    ...bigCookieClicks,
+    ...buildingsLevel,
+    ...buildingsOwned,
+    ...cookiesFromClicking,
+    ...everyBuildingOwned,
+    ...prestigeAmount,
+    ...prestigeUpgradesOwned,
+    ...shadow,
+    ...themedUpgrades,
+    ...totalBuildingsOwned,
+    ...totalUpgradesOwned,
+];
+
+export default checkHookContent;
