+++ src/Darkys_Achievement_Package/init/createHook/components/buildingsOwned/index.js
@@ -1,0 +1,59 @@
+import Achievement from "AchievementsRoot/Achievement";
+
+const buildingsOwned = () => {
+    Achievement("Squeak goes the mouse", "Have <b>900</b> cursors.", [0, 19]);
+    Game.last.order = 1060;
+
+    Achievement("I'm too old for this", "Have <b>650</b> grandmas.", [1, 21]);
+    Game.last.order = 1110;
+
+    Achievement("Gorden garden", "Have <b>650</b> farms.", [2, 21]);
+    Game.last.order = 1210;
+
+    Achievement("A fine mine you got there", "Have <b>650</b> mines.", [3, 21]);
+    Game.last.order = 1310;
+
+    Achievement("Working 24/7", "Have <b>650</b> factories.", [4, 21]);
+    Game.last.order = 1410;
+
+    Achievement("Bank on it", "Have <b>650</b> banks.", [15, 21]);
+    Game.last.order = 1430;
+
+    Achievement("Temple pun", "Have <b>650</b> temples.", [16, 21]);
+    Game.last.order = 1460;
+
+    Achievement("It's magic, Joel!", "Have <b>650</b> wizard towers.", [17, 21]);
+    Game.last.order = 1480;
+
+    Achievement("My shipment arrives in the morning", "Have <b>650</b> shipments.", [5, 21]);
+    Game.last.order = 1510;
+
+    Achievement("Turning cookies into gold since 2013", "Have <b>650</b> alchemy labs.", [6, 21]);
+    Game.last.order = 1610;
+
+    Achievement("A portal inside of a portal", "Have <b>650</b> portals.", [7, 21]);
+    Game.last.order = 1710;
+
+    Achievement("It all began on the 8th of August, 2013", "Have <b>650</b> time machines.", [8, 21]);
+    Game.last.order = 1810;
+
+    Achievement("AntimatterCon", "Have <b>650</b> antimatter condensers.", [13, 21]);
+    Game.last.order = 1910;
+
+    Achievement("Stop being a prissm", "Have <b>650</b> prisms.", [14, 21]);
+    Game.last.order = 2010;
+
+    Achievement("And yet you probably never got a blab", "Have <b>650</b> chancemakers.", [19, 21]);
+    Game.last.order = 2110;
+
+    Achievement("Gouchnox", "Have <b>650</b> fractal engines.<q>69</q>", [20, 21]);
+    Game.last.order = 2210;
+
+    Achievement("Follow the javascript", "Have <b>650</b> javascript consoles.", [32, 21]);
+    Game.last.order = 2310;
+
+    Achievement("Don't call them fanverses, you fool", "Have <b>650</b> idleverses.", [33, 21]);
+    Game.last.order = 2410;
+};
+
+export default buildingsOwned;
