+++ src/Darkys_Achievement_Package/init/createHook/components/cookiesFromClicking/index.js
@@ -1,0 +1,78 @@
+import Achievement from "AchievementsRoot/Achievement";
+
+const cookiesFromClicking = () => {
+    Achievement("Stop it, it clickles", "Make <b>100 octillion</b> cookies from clicking.", [11, 21]);
+    Game.last.order = 1010;
+
+    Achievement("Keep up the click", "Make <b>10 nonillion</b> cookies from clicking.", [
+        0,
+        69,
+        "https://i.imgur.com/3jNJJNw.png",
+    ]);
+    Game.last.order = 1011;
+
+    Achievement("Clicking is the new booping", "Make <b>1 decillion</b> cookies from clicking.", [
+        0,
+        34,
+        "https://i.imgur.com/3jNJJNw.png",
+    ]);
+    Game.last.order = 1012;
+
+    Achievement("The final click", "Make <b>100 decillion</b> cookies from clicking.<q>But its never over.</q>", [
+        0,
+        46,
+        "https://i.imgur.com/3jNJJNw.png",
+    ]);
+    Game.last.order = 1013;
+
+    Achievement(
+        "Sike, thats the wrong achievement",
+        "Make <b>10 undecillion</b> cookies from clicking.<q>The last achievement wasn't the final one.</q>",
+        [0, 5, "https://i.imgur.com/AWmeHiO.png"]
+    );
+    Game.last.order = 1014;
+
+    Achievement(
+        "How about we tap the cookie now",
+        "Make <b>1 duodecillion</b> cookies from clicking.<q>I heard the mobile version is out now.</q>",
+        [1, 5, "https://i.imgur.com/AWmeHiO.png"]
+    );
+    Game.last.order = 1015;
+
+    Achievement(
+        "Cookie Destroyer",
+        "Make <b>100 duodecillion</b> cookies from clicking.<q>Agent Johnson was here</q>",
+        [2, 5, "https://i.imgur.com/AWmeHiO.png"]
+    );
+    Game.last.order = 1016;
+
+    Achievement("Give it a click", "Make <b>10 tredecillion</b> cookies from clicking.", [
+        3,
+        5,
+        "https://i.imgur.com/AWmeHiO.png",
+    ]);
+    Game.last.order = 1017;
+
+    Achievement("Clicking Frenzy: Our Lord and Saviour", "Make <b>1 quattuordecillion</b> cookies from clicking.", [
+        4,
+        5,
+        "https://i.imgur.com/AWmeHiO.png",
+    ]);
+    Game.last.order = 1018;
+
+    Achievement("Hand-made cookies are still the best", "Make <b>100 quattuordecillion</b> cookies from clicking.", [
+        5,
+        5,
+        "https://i.imgur.com/AWmeHiO.png",
+    ]);
+    Game.last.order = 1019;
+
+    Achievement("My strange clicking addiction", "Make <b>10 quindecillion</b> cookies from clicking.", [
+        6,
+        5,
+        "https://i.imgur.com/AWmeHiO.png",
+    ]);
+    Game.last.order = 1019;
+};
+
+export default cookiesFromClicking;
