+++ src/Darkys_Achievement_Package/init/createHook/index.js
@@ -1,0 +1,31 @@
+import {
+    bigCookieClicks,
+    buildingsLevel,
+    buildingsOwned,
+    cookiesFromClicking,
+    etc,
+    everyBuildingOwned,
+    prestigeAmount,
+    prestigeUpgradesOwned,
+    shadow,
+    themedUpgrades,
+    totalBuildingsOwned,
+    totalUpgradesOwned,
+} from "./components";
+
+const createHook = () => {
+    bigCookieClicks();
+    buildingsLevel();
+    buildingsOwned();
+    cookiesFromClicking();
+    etc();
+    everyBuildingOwned();
+    prestigeAmount();
+    prestigeUpgradesOwned();
+    shadow();
+    themedUpgrades();
+    totalBuildingsOwned();
+    totalUpgradesOwned();
+};
+
+export default createHook;
