+++ src/Darkys_Achievement_Package/load/index.js
@@ -1,0 +1,9 @@
+const load = saveString => {
+    const save = JSON.parse(saveString);
+
+    save.unlockedAchievs.forEach(achievementName => {
+        Game.Achievements[achievementName].won = 1;
+    });
+};
+
+export default load;
