+++ src/Darkys_Armful_Collection_of_Upgrades/TieredUpgrade/index.js
@@ -1,0 +1,8 @@
+const TieredUpgrade = (name, desc, building, tier) => {
+    const answer = new Game.TieredUpgrade(name, desc, building, tier);
+    answer.darky = 1;
+
+    return answer;
+};
+
+export default TieredUpgrade;
