+++ src/Darkys_Armful_Collection_of_Upgrades/Upgrade/index.js
@@ -1,0 +1,8 @@
+const Upgrade = (name, desc, price, icon, buyFunction) => {
+    const answer = new Game.Upgrade(name, desc, price, icon, buyFunction);
+    answer.darky = 1;
+
+    return answer;
+};
+
+export default Upgrade;
