+++ src/Darkys_Armful_Collection_of_Upgrades/init/checkHook/components/index.js
@@ -1,0 +1,2 @@
+export { default as original } from "./original";
+export { default as tiered } from "./tiered";
