+++ src/Darkys_Armful_Collection_of_Upgrades/init/checkHook/components/original/index.js
@@ -1,0 +1,28 @@
+const original = [
+    // () => {
+    //     if (
+    //         Game.Objects["Cursor"].amount +
+    //             Game.Objects["Grandma"].amount +
+    //             Game.Objects["Farm"].amount +
+    //             Game.Objects["Factory"].amount +
+    //             Game.Objects["Mine"].amount +
+    //             Game.Objects["Bank"].amount +
+    //             Game.Objects["Temple"].amount +
+    //             Game.Objects["Wizard tower"].amount +
+    //             Game.Objects["Shipment"].amount +
+    //             Game.Objects["Alchemy lab"].amount +
+    //             Game.Objects["Portal"].amount +
+    //             Game.Objects["Time machine"].amount +
+    //             Game.Objects["Antimatter condenser"].amount +
+    //             Game.Objects["Prism"].amount +
+    //             Game.Objects["Chancemaker"].amount +
+    //             Game.Objects["Fractal engine"].amount +
+    //             Game.Objects["Javascript console"].amount >=
+    //         100
+    //     ) {
+    //         Game.Game.Unlock("Building up");
+    //     }
+    // },
+];
+
+export default original;
