+++ src/Darkys_Armful_Collection_of_Upgrades/init/checkHook/index.js
@@ -1,0 +1,5 @@
+import { original, tiered } from "./components";
+
+const checkHookContent = [...original, ...tiered];
+
+export default checkHookContent;
