+++ src/Darkys_Armful_Collection_of_Upgrades/init/createHook/components/original/index.js
@@ -1,0 +1,13 @@
+import Upgrade from "UpgradesRoot/Upgrade";
+
+const original = {
+    // No idea how to make this one work at all, help would be appreciated! The idea was that Everything is 0.005% more efficient per building.
+    // Upgrade(
+    //     "Building up",
+    //     "Everything is more efficient per building owned.<q>I can already feel the power building up.</q>",
+    //     1000000,
+    //     [2, 6]
+    // );
+};
+
+export default original;
