+++ src/Darkys_Armful_Collection_of_Upgrades/init/createHook/index.js
@@ -1,0 +1,8 @@
+import { original, tiered } from "./components";
+
+const createHook = () => {
+    original();
+    tiered();
+};
+
+export default createHook;
