+++ src/Darkys_Armful_Collection_of_Upgrades/init/index.js
@@ -1,0 +1,12 @@
+import checkHookContent from "./checkHook";
+import createHook from "./createHook";
+import overrides from "./overrides";
+
+const init = () => {
+    overrides();
+
+    Game.registerHook("check", checkHookContent);
+    createHook();
+};
+
+export default init;
